+++ src/ast/sub.rs
@@ -16,13 +16,13 @@ impl SubNode {
 }
 
 impl BinaryOpFunction for SubNode {
-    fn exec(&self, a: i32, b: i32) -> i32 {
+    fn exec(&self, a: f64, b: f64) -> f64 {
         a - b
     }
 }
 
 impl ASTNode for SubNode {
-    fn eval(&self, symtab: &mut HashMap<String, i32>) -> Result<i32, InterpreterError> {
+    fn eval(&self, symtab: &mut HashMap<String, f64>) -> Result<f64, InterpreterError> {
         let left = self.left.eval(symtab)?;
         let right = self.right.eval(symtab)?;
